post

+++ app/Http/Controllers/TasksController.php
@@ -16,12 +16,23 @@ class TasksController extends Controller
     public function index()
     {
          // タスク一覧を取得
-        $tasks = Task::all();
+        //$tasks = Task::all();
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスクの一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // 一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -52,11 +63,11 @@ class TasksController extends Controller
             'status'=>'required|max:10',
         ]);
         
-         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+         //  認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content'=>$request->content,
+            'status'=>$request->status,
+            ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -72,11 +83,31 @@ class TasksController extends Controller
     {
          // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-
+        
+         if (\Auth::id() === $task->user_id) {
+             
         // 詳細ビューでそれを表示
         return view('tasks.show', [
-            'task' => $task,
+           'task' => $task,
         ]);
+         }
+        else{
+          return redirect('/');
+        }
+        
+         // idの値でユーザを検索して取得
+      //  $user = User::findOrFail($id);
+
+         // ユーザの投稿一覧を作成日時の降順で取得
+      //  $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+     //   $data = [
+     //       'user'=>$user,
+     //       'tasks'=>$tasks,
+     //       ];
+            
+         // ユーザ詳細ビューでそれらを表示
+     //    return view('tasks.show',$data);
     }
 
     /**
@@ -89,11 +120,18 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) {        
 
         // メッセージ編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
+        
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -112,6 +150,7 @@ class TasksController extends Controller
         
          // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
         // タスクを更新
         $task->status = $request->status;
         $task->content = $request->content;
@@ -129,10 +168,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+         $task = \App\Task::find($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
